enable platformio
