+++ platformio.ini
@@ -1,0 +1,25 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:espwroom32]
+platform = espressif32
+board = lolin_d32
+framework = arduino
+build_flags = -D PIO_FRAMEWORK_ARDUINO_LWIP2_LOW_MEMORY -DLOGGING_REDEFINE_LOG_X
+board_build.partitions = min_spiffs.csv
+monitor_filters = colorize, esp32_exception_decoder
+monitor_port = /dev/cu.wchusbserial*
+upload_port = /dev/cu.wchusbserial*
+monitor_speed = 115200
+upload_speed = 921600
+lib_deps = 
+	bblanchon/ArduinoJson@^6.21.3
+	knolleary/PubSubClient@^2.8
+	https://github.com/esp32m/logging.git
